+++ src/retention/sqlite/statements.rs
@@ -16,10 +16,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{borrow::Cow, collections::HashSet, num::NonZeroUsize, time::Duration};
+use std::{borrow::Cow, collections::HashSet, time::Duration};
 
 use rusqlite::{Connection, OptionalExtension, Transaction};
-use tracing::{debug, instrument, trace, warn};
+use tracing::{instrument, trace, warn};
 
 use crate::{
     retention::{Id, PublishInfo, StoredInterface},
@@ -50,10 +50,9 @@ impl WriteConnection {
             }
         });
 
-        let capacity = self.store_capacity();
-        let transaction = self.transaction().map_err(SqliteError::Transaction)?;
+        self.free_retention_items(1)?;
 
-        Self::free_space(&transaction, capacity)?;
+        let transaction = self.transaction().map_err(SqliteError::Transaction)?;
 
         wrap_sync_call(|| {
             if !exists {
@@ -123,76 +122,10 @@ impl WriteConnection {
         Ok(())
     }
 
-    #[instrument(skip(self))]
-    pub(crate) async fn set_store_capacity(
-        &mut self,
-        store_capacity: NonZeroUsize,
-    ) -> Result<(), SqliteError> {
-        if store_capacity.get() == 0 {
-            warn!("cannot set store capacity to 0, maintaining the actual one");
-            return Ok(());
-        }
-
-        // before setting the new capacity, we perform a cleanup operation
-        let transaction = self.transaction().map_err(SqliteError::Transaction)?;
-        Self::free_space(&transaction, store_capacity.get())?;
-        transaction.commit().map_err(SqliteError::Transaction)?;
-
-        self.store_capacity = store_capacity;
-
-        Ok(())
-    }
-
-    /// Empty space when the database is full to allow storing newer elements
-    ///
-    /// In sequence
-    /// - first we check the occupied space is greater than the capacity
-    /// - if so, remove the expired elements from the store
-    /// - if the available space is still unsufficient, remove the oldest elements
-    #[instrument(skip_all)]
-    pub(crate) fn free_space(
-        transaction: &Transaction,
-        capacity: usize,
-    ) -> Result<(), SqliteError> {
-        debug!("remove elements from strore if it's full");
-
-        let count_stored = Self::count_stored(transaction)?;
-
-        trace!("initial count: {count_stored}");
-
-        if count_stored < capacity {
-            return Ok(());
-        }
-
-        trace!("store is full, evicting expired properties");
-        Self::delete_expired_transaction(transaction, &TimestampSecs::now())?;
-
-        // retrieve the updated number of stored properties after the removal of the expired ones
-        let count_stored = Self::count_stored(transaction)?;
-
-        trace!("count after expired removal: {count_stored}");
-
-        if count_stored < capacity {
-            return Ok(());
-        }
-
-        // TODO: decide whether to remove exactly the oldest elements or more
-        // this way the cleaning operation is ammortized over time
-        let to_remove = count_stored - capacity + 1;
-
-        trace!("store is still full, evicting the {to_remove} oldest elements");
-
-        let removed = Self::remove_oldest(transaction, to_remove)?;
-
-        debug!("removed {} elements", removed);
-
-        Ok(())
-    }
-
     /// Retrieve the number of stored properties
-    fn count_stored(transaction: &Transaction) -> Result<usize, SqliteError> {
+    pub(crate) fn count_stored(&self) -> Result<usize, SqliteError> {
         wrap_sync_call(|| {
-            let mut statement = transaction
+            let mut statement = self
                 .prepare_cached(include_query!("queries/retention/read/count_stored.sql"))
                 .map_err(SqliteError::Prepare)?;
 
@@ -203,13 +136,9 @@ impl WriteConnection {
     }
 
     /// Remove the N oldest elements from the store
-    fn remove_oldest(transaction: &Transaction, to_remove: usize) -> Result<usize, SqliteError> {
-        if to_remove == 0 {
-            return Ok(0);
-        }
-
+    pub(crate) fn remove_oldest(&self, to_remove: usize) -> Result<usize, SqliteError> {
         wrap_sync_call(|| {
-            let mut statement = transaction
+            let mut statement = self
                 .prepare_cached(include_query!(
                     "queries/retention/write/delete_n_oldest.sql"
                 ))
@@ -292,7 +221,7 @@ impl WriteConnection {
         Ok(())
     }
 
-    pub(super) fn delete_expired(&self, now: &TimestampSecs) -> Result<(), SqliteError> {
+    pub(super) fn delete_expired(&self, now: &TimestampSecs) -> Result<usize, SqliteError> {
         wrap_sync_call(|| {
             let mut statement = self
                 .prepare_cached(include_query!("queries/retention/write/delete_expired.sql"))
@@ -301,27 +230,7 @@ impl WriteConnection {
             let timestamp = now.to_bytes();
             let timestamp = timestamp.as_slice();
 
-            statement.execute([timestamp])?;
-
-            Ok(())
-        })
-    }
-
-    pub(super) fn delete_expired_transaction(
-        transaction: &Transaction,
-        now: &TimestampSecs,
-    ) -> Result<(), SqliteError> {
-        wrap_sync_call(|| {
-            let mut statement = transaction
-                .prepare_cached(include_query!("queries/retention/write/delete_expired.sql"))
-                .map_err(SqliteError::Prepare)?;
-
-            let timestamp = now.to_bytes();
-            let timestamp = timestamp.as_slice();
-
-            statement.execute([timestamp])?;
-
-            Ok(())
+            statement.execute([timestamp]).map_err(SqliteError::Query)
         })
     }
 
@@ -457,7 +366,9 @@ pub(crate) mod tests {
     use itertools::Itertools;
 
     use crate::{
-        retention::{Context, StoredRetention, TimestampMillis},
+        retention::{
+            sqlite::tests::publish_with_expiry, Context, StoredRetention, TimestampMillis,
+        },
         store::SqliteStore,
     };
 
@@ -1177,5 +1088,80 @@ pub(crate) mod tests {
         });
 
         assert_eq!(expected, res);
+    }
+
+    #[tokio::test]
+    async fn should_count_stored() {
+        let dir = tempfile::tempdir().unwrap();
+
+        let store = SqliteStore::connect(dir.path()).await.unwrap();
+
+        let id1 = Context::new().next();
+        store
+            .store_publish(&id1, publish_with_expiry("/path1", None))
+            .await
+            .unwrap();
+
+        let count = {
+            let connection = store.writer.lock().await;
+            connection.count_stored().unwrap()
+        };
+
+        assert_eq!(count, 1);
+    }
+
+    #[tokio::test]
+    async fn should_remove_oldest() {
+        let dir = tempfile::tempdir().unwrap();
+
+        let store = SqliteStore::connect(dir.path()).await.unwrap();
+
+        // insert 3 elements and check removal properly works
+        let ctx = Context::new();
+
+        let id1 = ctx.next();
+        store
+            .store_publish(&id1, publish_with_expiry("/path1", None))
+            .await
+            .unwrap();
+
+        let id2 = ctx.next();
+        store
+            .store_publish(&id2, publish_with_expiry("/path2", None))
+            .await
+            .unwrap();
+
+        let id3 = ctx.next();
+        store
+            .store_publish(&id3, publish_with_expiry("/path3", None))
+            .await
+            .unwrap();
+
+        let connection = store.writer.lock().await;
+
+        let removed = connection.remove_oldest(0).unwrap();
+        assert_eq!(removed, 0);
+
+        let removed = connection.remove_oldest(2).unwrap();
+        assert_eq!(removed, 2);
+
+        // check that the remaining publish is the last one
+        let res = fetch_publish(&store, &id3).unwrap();
+
+        let publish3 = RetentionPublish {
+            id: id3,
+            interface: "com.Foo".into(),
+            path: "/path3".into(),
+            payload: [].as_slice().into(),
+            sent: false,
+            expiry_time: None,
+        };
+
+        assert_eq!(res, publish3);
+
+        // try removing more elements than available
+        let removed = connection.remove_oldest(2).unwrap();
+        // only 1 element remains from the previous removal
+        assert_eq!(removed, 1);
     }
 }